sprint planning (#58)
* sprint planning

* refactor to support bff

* more work on plan view bff code

* refactor and use sprints from database

* move StateTree type + fix sprint status issue

* implemented points info text

* drop debugger statement

* a little refactoring and fixing build issues

* fix date range display bug

* invertable icon

* collapse icon in place

* expand and collapse sprint panels

* added sprint backlog item resource

* sprint backlog items ui

* break up BacklogItemPlanningPanel to smaller files

* refactored BacklogItemPlanningPanel some more

* UI for adding sprint backlog items

* checkbox component implemented

* checkboxes to select backlog items

* fixing sprint planning panel layout for mobile

* fixed some styling issues + content

* fix desktop styling issue

* fix button alignment on mobile issue

* ensure mobile checkbox placed correctly

* selected items + add items button state

* fix cursor css issue

* api batch almost working

* got api batch working, just ui work remaining

* backlog item to sprint item

* handle failed api call

* improve interfaces and clean up

* changes to pop detail menu for sprint backlog

* split out item menu & fixed add pbi bug

* refactor out item menu builder

* dispatch move to product backlog

* got api call working

* got remove from sprint backlog working

* added constant for mobile css class

* buttons for adding sprints before & after

* rename actions etc. to match ui

* one missing change

* add:sprint action with data - no UI yet

* sort sprints

* initial work for sprint edit form

* add new sprint panel styling complete

* fix for mobile sprint detail form

* styling for mobile

* abilty to edit sprint name field + others

* implemented sprint cancel button

* added POST sprint handler

* add sprint working (not refreshing UI yet)

* added sprint api actions and refactor

* remove unnecessary sprint.displayindex field

* clean up

* defensive code to handle null backlog items

* abbreviate month name

* clean up

* fix sprint planning panel storybook issue

* fixes for tests etc.

* refactor to creatre sprint card

* SprintPlanningPanelSprint to SprintCardSprint

* item detail button for sprint

* state update to show sprint item menu

* sprint item detail menu styling

* handle click outside menu

* fix sprint backlog item detail menu issue

* handle app click to hide sprint backlog detail menu

* fix gap between sprint backlog items

* tweaks for item detail menu

* delete sprint api action should work now...

* future use note

* removed TODOs and added issue 161 for this

* type defn updates

* removed TODO

* dealt with TODO

* remove another todo

* organize imports

* switch comment to TODO

* remove unneeded param

* remove unused file

* remove todo and added issue 162

* organize imports

* explain commented code

* organize imports

* organize imports

* clean up

+++ src/components/organisms/combo/BacklogItemPlanningItem.tsx
@@ -3,12 +3,16 @@ import * as React from "react";
 import { useDispatch } from "react-redux";
 
 // interfaces/types
-import { BacklogItemWithSource } from "../../../reducers/backlogItemsReducer";
+import { BacklogItemWithSource } from "../../../reducers/backlogItems/backlogItemsReducerTypes";
 import { EditMode } from "../../molecules/buttons/EditButton";
 import { SimpleDivider } from "../../atoms/dividers/SimpleDivider";
 import { BacklogItemDetailForm } from "../forms/BacklogItemDetailForm";
-import { BacklogItemCard, BacklogItemTypeEnum } from "../../molecules/cards/BacklogItemCard";
+import { BacklogItemCard, BacklogItemTypeEnum, ItemMenuEventHandlers } from "../../molecules/cards/BacklogItemCard";
+
+// utils
+import { calcItemId } from "../../molecules/cards/BacklogItemCard";
 import { buildClassName } from "../../../utils/classNameBuilder";
+import { productBacklogItemMenuBuilder } from "../../common/itemMenuBuilders";
 
 // actions
 import { apiDeleteBacklogItem } from "../../../actions/apiBacklogItems";
@@ -19,8 +23,10 @@ import {
     backlogItemDetailClicked,
     editBacklogItem,
     cancelEditBacklogItem,
-    updateBacklogItem
-} from "../../../actions/backlogItems";
+    updateBacklogItem,
+    selectProductBacklogItem,
+    unselectProductBacklogItem
+} from "../../../actions/backlogItemActions";
 
 // style
 import css from "./BacklogItemPlanningItem.module.css";
@@ -84,32 +90,46 @@ export const BacklogItemPlanningItem: React.FC<BacklogItemPlanningItemProps> = (
             </>
         );
     } else {
+        const itemEventHandlers: ItemMenuEventHandlers = {
+            handleEvent: (eventName: string, itemId: string) => {
+                if (eventName === "onEditItemClicked") {
+                    dispatch(editBacklogItem(props.id));
+                } else if (eventName === "onRemoveItemClicked") {
+                    dispatch(apiDeleteBacklogItem(props.id));
+                } else {
+                    throw Error(`${eventName} is not handled`);
+                }
+            }
+        };
         return (
             <>
                 <SimpleDivider key={`divider-saved-${props.id}`} hidden={props.hidden} highlighted={props.highlightAbove} />
                 <BacklogItemCard
-                    key={props.id}
+                    buildItemMenu={productBacklogItemMenuBuilder(itemEventHandlers)}
                     estimate={props.estimate}
+                    hasDetails={props.editMode === EditMode.Edit}
                     hidden={props.hidden}
                     internalId={`${props.id}`}
-                    itemId={`${props.externalId || props.friendlyId}`}
-                    itemType={props.type === "story" ? BacklogItemTypeEnum.Story : BacklogItemTypeEnum.Bug}
-                    titleText={props.storyPhrase}
                     isDraggable={props.editMode === EditMode.Edit}
-                    hasDetails={props.editMode === EditMode.Edit}
-                    renderMobile={props.renderMobile}
+                    isSelectable={props.editMode === EditMode.Edit}
+                    itemId={calcItemId(props.externalId, props.friendlyId)}
+                    itemType={props.type === "story" ? BacklogItemTypeEnum.Story : BacklogItemTypeEnum.Bug}
+                    key={props.id}
                     marginBelowItem
                     pushState={props.pushState}
+                    renderMobile={props.renderMobile}
+                    showDetailMenu={props.showDetailMenu}
+                    titleText={props.storyPhrase}
                     onDetailClicked={() => {
                         dispatch(backlogItemDetailClicked(props.id));
                     }}
-                    onRemoveItemClicked={(backlogItemId) => {
-                        dispatch(apiDeleteBacklogItem(props.id));
+                    onCheckboxChange={(checked) => {
+                        if (checked) {
+                            dispatch(selectProductBacklogItem(props.id));
+                        } else {
+                            dispatch(unselectProductBacklogItem(props.id));
+                        }
                     }}
-                    onEditItemClicked={(itemId: string) => {
-                        dispatch(editBacklogItem(props.id));
-                    }}
-                    showDetailMenu={props.showDetailMenu}
                 />
             </>
         );
